feat: add context extenders classes
