+++ common/oatbox/log/ServiceProvider/LogServiceProvider.php
@@ -24,6 +24,9 @@ namespace oat\oatbox\log\ServiceProvider;
 
 use oat\generis\model\DependencyInjection\ContainerServiceProviderInterface;
 use oat\oatbox\log\logger\AdvancedLogger;
+use oat\oatbox\log\logger\extender\ExceptionContextExtender;
+use oat\oatbox\log\logger\extender\RequestContextExtender;
+use oat\oatbox\log\logger\extender\UserContextExtender;
 use oat\oatbox\log\LoggerService;
 use oat\oatbox\session\SessionService;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
@@ -36,12 +39,38 @@ class LogServiceProvider implements ContainerServiceProviderInterface
     {
         $services = $configurator->services();
 
+        $services->set(ExceptionContextExtender::class, ExceptionContextExtender::class);
+        $services->set(RequestContextExtender::class, RequestContextExtender::class);
+        $services->set(UserContextExtender::class, UserContextExtender::class)
+            ->args(
+                [
+                    service(SessionService::SERVICE_ID),
+                ]
+            );
+
         $services->set(AdvancedLogger::class, AdvancedLogger::class)
             ->public()
+            ->call(
+                'addContextExtender',
+                [
+                    service(ExceptionContextExtender::class),
+                ]
+            )
+            ->call(
+                'addContextExtender',
+                [
+                    service(RequestContextExtender::class),
+                ]
+            )
+            ->call(
+                'addContextExtender',
+                [
+                    service(UserContextExtender::class),
+                ]
+            )
             ->args(
                 [
                     service(LoggerService::SERVICE_ID),
-                    service(SessionService::SERVICE_ID),
                 ]
             );
     }
